% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{choose_dark_or_light}
\alias{choose_dark_or_light}
\title{Choose dark or light color}
\usage{
choose_dark_or_light(x, black = darken(x, 0.8), white = lighten(x,
  0.8))
}
\arguments{
\item{x}{The background color}

\item{black}{Text or foreground color, e.g. "#22222" or \code{darken(x, 0.8)}, if
black text provides the best contrast.}

\item{white}{Text or foreground color or expression, e.g. "#EEEEEE" or
\code{lighten(x, 0.8)}, if white text provides the best contrast.}
}
\description{
Takes a color input as \code{x} and returns either the black or white color if
dark or light text should be used over the input color for best contrast.
Follows W3C Recommendations.
}
\references{
\url{https://stackoverflow.com/a/3943023/2022615}
}
