% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{countdown_shiny_example}
\alias{countdown_shiny_example}
\title{Example Countdown Shiny App}
\usage{
countdown_shiny_example(display.mode = c("showcase", "normal", "auto"))
}
\arguments{
\item{display.mode}{The mode in which to display the application. If set to
the value \code{"showcase"}, shows application code and metadata from a
\code{DESCRIPTION} file in the application directory alongside the
application. If set to \code{"normal"}, displays the application normally.
Defaults to \code{"auto"}, which displays the application in the mode given
in its \code{DESCRIPTION} file, if any.}
}
\value{
Runs the example Shiny app in the current R session.
}
\description{
An example app that demonstrates the ways that countdown timers can be
integrated into Shiny apps.
}
\examples{
if (interactive()) {
  countdown_shiny_example()
}

}
\seealso{
Other Shiny functions: 
\code{\link{countdown_action}()},
\code{\link{countdown_app}()},
\code{\link{countdown_update}()}
}
\concept{Shiny functions}
