% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countdown.R
\name{countdown_timer}
\alias{countdown_timer}
\title{Countdown Timer
Inserts a countdown timer into R Markdown documents}
\usage{
countdown_timer(seconds)
}
\arguments{
\item{seconds}{The number of seconds to count down}
}
\description{
Countdown Timer
Inserts a countdown timer into R Markdown documents
}

