% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countdown-package.R
\docType{package}
\name{countdown-package}
\alias{countdown-package}
\alias{_PACKAGE}
\title{countdown: A Countdown Timer for HTML Presentations, Documents, and Web Apps}
\description{
A simple countdown timer for slides and HTML documents written in 'R Markdown' or 'Quarto'. Integrates fully into 'Shiny' apps. Countdown to something amazing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.garrickadenbuie.com/countdown/}
  \item \url{https://github.com/gadenbuie/countdown}
  \item Report bugs at \url{https://github.com/gadenbuie/countdown/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID}) [copyright holder]

}
\keyword{internal}
