% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countdown.R
\name{countdown}
\alias{countdown}
\title{Countdown Widget}
\usage{
countdown(
  count,
  duration = 2500,
  format = "(,ddd).dd",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{count}{Value to animate}

\item{duration}{Milliseconds}

\item{format}{Value format}

\item{width, height}{Dimensions}

\item{elementId}{Id of element}
}
\description{
Odmoter JS R Wrapper
}
\examples{
countdown(347)

}
