% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{countdown_app}
\alias{countdown_app}
\title{Launch Countdown Shiny App}
\usage{
countdown_app(...)
}
\arguments{
\item{...}{Arguments passed on to \code{shiny::runApp}
\describe{
  \item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}
  \item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started. Defaults to true in
interactive sessions only. This value of this parameter can also be a
function to call with the application's URL.}
  \item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}
  \item{workerId}{Can generally be ignored. Exists to help some editions of
Shiny Server Pro route requests to the correct process.}
  \item{quiet}{Should Shiny status messages be shown? Defaults to FALSE.}
  \item{test.mode}{Should the application be launched in test mode? This is
only used for recording or running automated tests. Defaults to the
\code{shiny.testmode} option, or FALSE if the option is not set.}
}}
}
\description{
Launches a full screen, interactive countdown timer as a \link{shiny} app.
}
